# router/internal/handlers/driver.go
// Handlers for driver endpoints
package handlers

import (
	"net/http"
	"time"

	"github.com/gin-gonic/gin"

	"github.com/Macklin06/optiroute/router/internal/models"
	"github.com/Macklin06/optiroute/router/internal/services"
)

// DriverHandler holds the service layer reference
type DriverHandler struct {
	DriverService *services.DriverService
}

// NewDriverHandler creates a new handler
func NewDriverHandler(driverService *services.DriverService) *DriverHandler {
	return &DriverHandler{
		DriverService: driverService,
	}
}

// UpdateLocation godoc
// @Summary Update driver location
// @Description Updates the live location of a driver and stores it in Redis and PostgreSQL
// @Tags Drivers
// @Accept json
// @Produce json
// @Param request body models.LocationUpdate true "Driver location update"
// @Success 200 {object} map[string]interface{}
// @Failure 422 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /api/v1/drivers/location [put]
// UpdateLocation updates a driver's location (PUT /api/v1/drivers/location)
func (h *DriverHandler) UpdateLocation(c *gin.Context) {
	start := time.Now()
	status := "success"
	defer func() {
		duration := time.Since(start).Seconds()
		services.LocationUpdateDuration.WithLabelValues(status).Observe(duration)
		services.LocationUpdateTotal.WithLabelValues(status).Inc()
	}()

	// Parse and validate JSON request
	var req models.LocationUpdate

	if err := c.ShouldBindJSON(&req); err != nil {
		// Invalid request
		status = "validation_error"
		c.JSON(http.StatusUnprocessableEntity, gin.H{
			"error":   "invalid request body",
			"details": err.Error(),
		})
		return
	}

	// Call service to update location
	if err := h.DriverService.UpdateDriverLocation(req); err != nil {
		// Server error
		status = "error"
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": err.Error(),
		})
		return
	}

	// Success
	c.JSON(http.StatusOK, gin.H{
		"driver_id": req.DriverID,
		"latitude":  req.Latitude,
		"longitude": req.Longitude,
		"message":   "location updated successfully",
	})
}

// GetActiveDrivers godoc
// @Summary Get active drivers
// @Description Returns all currently active drivers from Redis cache
// @Tags Drivers
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /api/v1/drivers/active [get]
// GetActiveDrivers retrieves all active drivers from cache (GET /api/v1/drivers/active)
func (h *DriverHandler) GetActiveDrivers(c *gin.Context) {
	drivers, err := h.DriverService.GetActiveDrivers()
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": err.Error(),
		})
		return
	}

	services.ActiveDriversGauge.Set(float64(len(drivers)))

	c.JSON(http.StatusOK, gin.H{
		"drivers": drivers,
		"count":   len(drivers),
	})
}